@@ -10,16 +10,14 @@
         "predeploy": "next build && next start"
     },
     "dependencies": {
-        "@next/font": "13.1.6",
         "@types/marked": "^4.0.8",
         "@types/node": "18.11.18",
         "@types/react": "18.0.27",
         "@types/react-dom": "18.0.10",
         "@types/react-syntax-highlighter": "^15.5.6",
-        "@types/styled-components": "^5.1.26",
         "gray-matter": "^4.0.3",
         "marked": "^5.0.0",
-        "next": "13.1.6",
+        "next": "13.4.2",
         "react": "18.2.0",
         "react-dom": "18.2.0",
         "react-markdown": "^8.0.5",
@@ -27,7 +25,7 @@
         "react-syntax-highlighter": "^15.5.0",
         "rehype-raw": "^6.1.1",
         "sharp": "^0.31.3",
-        "styled-components": "^5.3.6",
+        "styled-components": "^6.0.0-rc.1",
         "typeit-react": "^2.6.4",
         "typescript": "4.9.5"
     },
