@@ -1,23 +1,18 @@
-import { GetStaticProps, GetStaticPropsContext } from 'next';
-import ReactMarkdown from 'react-markdown';
-import matter from 'gray-matter';
-import rehypeRaw from 'rehype-raw';
-import styled from 'styled-components';
-import fs from 'fs';
-import path from 'path';
-import { PostData } from '@/types/post';
+'use client';
+import ImageBox from '@/components/ImageBox';
 import BlogPostTitle from '@/components/PostTitle';
-import { notoSansKR } from '@/styles/fonts/notoSansKR';
-import { marked } from 'marked';
+import SideMenu from '@/components/SideMenu';
 import TOC from '@/components/TOC';
-import { Heading } from '@/types/heading';
-import Head from 'next/head';
+import ReactMarkdown from 'react-markdown';
+import rehypeRaw from 'rehype-raw';
+import Image from 'next/image';
 import CodeBlock from '@/components/CodeBlock';
 import { Prism as SyntaxHighlighter } from 'react-syntax-highlighter';
 import { oneDark } from 'react-syntax-highlighter/dist/cjs/styles/prism';
-import Image from 'next/image';
-import ImageBox from '@/components/ImageBox';
-import SideMenu from '@/components/SideMenu';
+import { styled } from 'styled-components';
+import { notoSansKR } from '@/styles/fonts/notoSansKR';
+import { PostData } from '@/types/post';
+import { Heading } from '@/types/heading';
 
 interface Props {
     postData: PostData;
@@ -25,20 +20,10 @@ interface Props {
     headings: Heading[];
 }
 
-const Post = (props: Props) => {
+const Main = (props: Props) => {
     const { postData, postContent, headings } = props;
-
     return (
         <>
-            <Head>
-                <title>{postData.title}</title>
-                <meta
-                    property="og:title"
-                    content={postData.title ? postData.title : 'default title'}
-                />
-                <meta property="og:image" content={postData.thumbnail} />
-                <meta property="og:description" content={postData.summary} />
-            </Head>
             <SideMenu />
             <Container>
                 <BlogPostTitle postData={postData}></BlogPostTitle>
@@ -138,50 +123,7 @@ const Post = (props: Props) => {
     );
 };
 
-export default Post;
-
-export const getStaticPaths = async () => {
-    return {
-        paths: [],
-        fallback: 'blocking',
-    };
-};
-
-export const getStaticProps: GetStaticProps = async (
-    context: GetStaticPropsContext,
-) => {
-    const postsDirectory = path.join(process.cwd(), 'src/assets/md');
-    const fullPath = path.join(postsDirectory, `${context.params?.slug}.md`);
-
-    if (!fs.existsSync(fullPath)) {
-        return {
-            notFound: true,
-        };
-    }
-
-    const rawMD = fs.readFileSync(fullPath, 'utf8');
-    const processedMD = matter(rawMD);
-
-    const headings: Heading[] = [];
-    marked
-        .parse(processedMD.content)
-        .split('\n')
-        .forEach(str => {
-            const match = str.match(
-                /<(h[1-3])\s+[^>]*id="([^"]+)"[^>]*>([^<]+)<\/\1>/,
-            );
-            if (match)
-                headings.push({ tag: match[1], id: match[2], value: match[3] });
-        });
-
-    return {
-        props: {
-            postData: processedMD.data,
-            postContent: processedMD.content,
-            headings,
-        },
-    };
-};
+export default Main;
 
 const Container = styled.div`
     width: 40%;
