@@ -1,4 +1,4 @@
-import { Noto_Sans_KR } from '@next/font/google';
+import { Noto_Sans_KR } from 'next/font/google';
 
 export const notoSansKR = Noto_Sans_KR({
     weight: '400',
